Move measurement timescales to their own file
No changes. Just moved the struct and test to a new file and replaced
includes where necessary
