+++ src/ControlSystem/Tags/CMakeLists.txt
@@ -6,4 +6,5 @@ spectre_target_headers(
   INCLUDE_DIRECTORY ${CMAKE_SOURCE_DIR}/src
   HEADERS
   FunctionsOfTimeInitialize.hpp
+  MeasurementTimescales.hpp
   )
