+++ tests/Unit/ControlSystem/Tags/CMakeLists.txt
@@ -4,4 +4,5 @@
 set(LIBRARY_SOURCES
   ${LIBRARY_SOURCES}
   Tags/Test_FunctionsOfTimeInitialize.cpp
+  Tags/Test_MeasurementTimescales.cpp
   PARENT_SCOPE)
